Adding spvGroupNonUniform to Slang compiler
Script sync from SHA: 60a55366682f9d441e966f3aca0e0e5e6399b07a

+++ raytrace_tutorial/11_shader_execution_reorder/11_shader_execution_reorder.cpp
@@ -410,8 +410,8 @@ public:
 
   void updateFrame()
   {
-    static glm::mat4 refCamMatrix;
-    static float     refFov{m_cameraManip->getFov()};
+    static glm::dmat4 refCamMatrix;
+    static double     refFov{m_cameraManip->getFov()};
 
     const auto& m   = m_cameraManip->getViewMatrix();
     const auto  fov = m_cameraManip->getFov();
