+++ raytrace_tutorial/16_ray_query/16_ray_query.cpp
@@ -265,8 +265,8 @@ public:
   // This enables progressive rendering where each frame accumulates more samples
   void updateFrame()
   {
-    static glm::mat4 refCamMatrix;
-    static float     refFov{m_cameraManip->getFov()};
+    static glm::dmat4 refCamMatrix;
+    static double     refFov{m_cameraManip->getFov()};
 
     const auto& m   = m_cameraManip->getViewMatrix();
     const auto  fov = m_cameraManip->getFov();
